Split C++ atomistic code into system and models
The files where becoming a too large
